+++ core/initializer/test_suite_runner.go
@@ -9,20 +9,8 @@ import (
 	"github.com/kurtosis-tech/kurtosis/initializer/parallelism"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
-	"io"
-	"io/ioutil"
 	"math"
 	"net"
-	"os"
-	"sort"
-)
-
-// =============================== "enum" for test result =========================================
-type testResult string
-const (
-	PASSED  testResult = "PASSED"
-	FAILED  testResult = "FAILED"
-	ERRORED testResult = "ERRORED" // Indicates an error during setup that prevented the test from running
 )
 
 // =============================== Test Suite Runner =========================================
@@ -137,16 +125,7 @@ func (runner TestSuiteRunner) RunTests(testNamesToRun map[string]bool, testParal
 		testParallelism)
 
 	logrus.Infof("Running %v tests with execution ID %v...", len(testsToRun), executionInstanceId.String())
-	interceptor := parallelism.NewErroneousSystemLogCaptureWriter()
-	testOutputs := testExecutor.RunInParallel(interceptor, testParams)
-
-	logrus.Infof("Printing results for %v tests...", len(testsToRun))
-	allTestsPassed = processTestOutputs(testsToRun, testOutputs)
-
-	// If there was any erroneous system-level logging during parallel test execution, loudly display that to the user
-	capturedErroneousMessages := interceptor.GetCapturedMessages()
-	logErroneousSystemLogging(capturedErroneousMessages)
-
+	allTestsPassed = testExecutor.RunInParallelAndPrintResults(testParams)
 	return allTestsPassed, nil
 }
 
@@ -183,150 +162,8 @@ func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]testsu
 		binary.BigEndian.PutUint32(subnetIp, subnetIpInt)
 		subnetCidrStr := fmt.Sprintf("%v/%v", subnetIp.String(), subnetMaskBits)
 
-		tempFilename := fmt.Sprintf("%v-%v", executionInstanceId, testName)
-		tempFp, err := ioutil.TempFile("", tempFilename)
-		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred creating temporary file to contain logs of test %v", testName)
-		}
-		logrus.Tracef("Temp file to write logs for test %v to: %v", testName, tempFp.Name())
-
-		testParams[testName] = *parallelism.NewParallelTestParams(testName, test, tempFp, subnetCidrStr, executionInstanceId)
+		testParams[testName] = *parallelism.NewParallelTestParams(testName, test, subnetCidrStr, executionInstanceId)
 		testIndex++
 	}
 	return testParams, nil
-}
-
-/*
-Helper function to process the TestExecutorParallelizer's output to print the necessary information to STDOUT
-
-Args:
-	testsToRun: A "set" of tests that were run
-	testOutputs: The output of the
-
-Returns:
-	True if all tests passed, false otherwise
- */
-func processTestOutputs(testsToRun map[string]testsuite.Test, testOutputs map[string]parallelism.ParallelTestOutput) bool {
-	// We want normalized output between runs of the tests suite so we sort the tests by name
-	testPrintOrder := []string{}
-	for testName, _ := range testsToRun {
-		testPrintOrder = append(testPrintOrder, testName)
-	}
-	sort.Strings(testPrintOrder)
-
-	allTestResults := make(map[string]testResult)
-	for _, name := range testPrintOrder {
-		printBanner(name, false)
-
-		output := testOutputs[name]
-		passed := output.TestPassed
-		executionErr := output.ExecutionErr
-		logFp := output.LogFp
-
-		// Close our log FP now that we're done writing, to switch to read mode
-		logFp.Close()
-		readLogFp, err := os.Open(logFp.Name())
-		if err != nil {
-			logrus.Error("An error occurred opening the test's logfile for reading; logs for this test are unavailable")
-			fmt.Fprintln(logrus.StandardLogger().Out, err) // Logrus will escape newlines so we don't actually log this
-		} else {
-			bytesWritten, err := io.Copy(logrus.StandardLogger().Out, readLogFp)
-			logrus.Tracef("Wrote %v bytes to STDOUT from test logfile", bytesWritten)
-			if err != nil {
-				logrus.Error("An error occurred copying the test's logfile to STDOUT; the logs above may not be complete!")
-				fmt.Fprintln(logrus.StandardLogger().Out, err) // Logrus will escape newlines so we don't actually log this
-			}
-		}
-		readLogFp.Close()
-		os.Remove(logFp.Name()) // We're responsible for cleaning up the temp file we created
-
-		result := logTestResult(name, executionErr, passed)
-		allTestResults[name] = result
-	}
-
-	printBanner("All Test Results", false)
-	allTestsPassed := true
-	for _, testName := range testPrintOrder {
-		result := allTestResults[testName]
-		logrus.Infof("- %v: %v", testName, result)
-		allTestsPassed = allTestsPassed && result == PASSED
-	}
-
-	return allTestsPassed
-}
-
-/*
-Handles determining the proper test status and logging it.
-Returns the testResult for convenience.
-*/
-func logTestResult(testName string, executionErr error, testPassed bool) testResult {
-	var result testResult
-	if executionErr != nil {
-		result = ERRORED
-	} else {
-		if testPassed {
-			result = PASSED
-		} else {
-			result = FAILED
-		}
-	}
-
-	switch result {
-	case ERRORED:
-		logrus.Errorf("Test %v %v", testName, result)
-		logrus.Errorf("Error reason: %v", executionErr)
-	case PASSED:
-		logrus.Infof("Test %v %v", testName, result)
-	case FAILED:
-		logrus.Errorf("Test %v %v", testName, result)
-	}
-	return result
-}
-
-func printBanner(contents string, isError bool) {
-	bannerString := "=================================================================================================="
-	contentString := fmt.Sprintf("                              %v", contents)
-	if !isError {
-		logrus.Info("")
-		logrus.Info(bannerString)
-		logrus.Info(contentString)
-		logrus.Info(bannerString)
-	} else {
-		logrus.Error("")
-		logrus.Error(bannerString)
-		logrus.Error(contentString)
-		logrus.Error(bannerString)
-	}
-}
-
-/*
-Helper function to print a big warning if there was logging to the system-level logging when there should only have been
- logging to the test-specific logger
- */
-func logErroneousSystemLogging(capturedErroneousMessages []parallelism.ErroneousSystemLogInfo) {
-	if len(capturedErroneousMessages) == 0 {
-		return
-	}
-
-	printBanner("Erroneous Logs", true)
-	logrus.Error("There were log messages printed to the system-level logger during parallel test execution!")
-	logrus.Error("Because the system-level logger is shared and the tests run in parallel, the messages cannot be")
-	logrus.Error(" attributed to any specific test. This is:")
-	logrus.Error("   1) A bug in Kurtosis, and a system-level logger call was used when a test-specific logger")
-	logrus.Error("       should have been used (likely)")
-	logrus.Error("   2) Third-party code calling logrus independently, and there's nothing we can do (unlikely, but possible)")
-	logrus.Error("")
-	logrus.Error("The log message(s) attempted, and the stacktrace(s) of origination, are as follows in the order they were logged:")
-	logrus.Error("")
-
-	for i, messageInfo := range capturedErroneousMessages {
-		logrus.Errorf("----------------- Erroneous Message #%d -------------------", i+1)
-		logrus.Error("Message:")
-		logrus.StandardLogger().Out.Write(messageInfo.Message)
-		logrus.StandardLogger().Out.Write([]byte("\n")) // The message likely won't come with a newline so we add it
-		logrus.Error("")
-		logrus.Error("Stacktrace:")
-		logrus.StandardLogger().Out.Write(messageInfo.Stacktrace)
-		logrus.StandardLogger().Out.Write([]byte("\n")) // The stacktrace likely won't end with a newline so we add it
-	}
 }